# go-backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  entities.ApiResponse:
    properties:
      data: {}
      message:
        type: string
      pagination:
        $ref: '#/definitions/entities.PaginationResponse'
      success:
        type: boolean
    type: object
...
  entities.Category:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      image:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  entities.CreateOrderRequest:
    properties:
      notes:
        type: string
      payment_method:
        type: string
      shipping_address:
        type: string
      shipping_method:
        type: string
    required:
    - payment_method
    - shipping_address
    - shipping_method
    type: object
  entities.CreatePaymentRequest:
    properties:
      order_id:
        type: string
      payment_data:
        type: string
      payment_method:
        type: string
    required:
    - order_id
    - payment_method
    type: object
...
  entities.Order:
    properties:
      created_at:
        type: string
      id:
        type: string
      notes:
        type: string
      order_items:
        items:
          $ref: '#/definitions/entities.OrderItem'
        type: array
      payment_method:
        type: string
      payment_status:
        type: string
      shipping_address:
        type: string
      shipping_method:
        type: string
      shipping_status:
        type: string
      status:
        type: string
      total_price:
        type: number
      tracking_number:
        type: string
      transactions:
        items:
          $ref: '#/definitions/entities.Transaction'
        type: array
      updated_at:
        type: string
      user:
        $ref: '#/definitions/entities.User'
      user_id:
        type: string
    type: object
  entities.OrderItem:
    properties:
      created_at:
        type: string
      id:
        type: string
      order_id:
        type: string
      price:
        type: number
      product:
        $ref: '#/definitions/entities.Product'
      product_id:
        type: string
      quantity:
        type: integer
      updated_at:
        type: string
    type: object
  entities.PaginationResponse:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total_items:
        type: integer
      total_pages:
        type: integer
    type: object
  entities.Permission:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  entities.Product:
    properties:
      category:
        $ref: '#/definitions/entities.Category'
      category_id:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      image:
        type: string
      images:
        items:
          $ref: '#/definitions/entities.ProductImage'
        type: array
      name:
        type: string
      price:
        type: number
      stock:
        type: integer
      updated_at:
        type: string
    type: object
  entities.ProductImage:
    properties:
      created_at:
        type: string
      id:
        type: string
      image_url:
        type: string
      product_id:
        type: string
      updated_at:
        type: string
    type: object
...
  entities.Role:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      permissions:
        items:
          $ref: '#/definitions/entities.Permission'
        type: array
      updated_at:
        type: string
    type: object
...
  entities.Transaction:
    properties:
      amount:
        type: number
      created_at:
        type: string
      id:
        type: string
      order_id:
        type: string
      payment_data:
        type: string
      payment_method:
        type: string
      status:
        type: string
      transaction_id:
        type: string
      updated_at:
        type: string
    type: object
...
  entities.UpdateOrderStatusRequest:
    properties:
      status:
        type: string
    required:
    - status
    type: object
...
  entities.User:
    properties:
      active:
        type: boolean
      address:
        type: string
      avatar:
        type: string
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: string
      last_name:
        type: string
      phone:
        type: string
      role:
        $ref: '#/definitions/entities.Role'
      role_id:
        type: string
      updated_at:
        type: string
    type: object
...
  entities.VerifyPaymentRequest:
    properties:
      payment_data:
        type: string
      transaction_id:
        type: string
    required:
    - transaction_id
    type: object
host: localhost:4001
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.swagger.io/support
  description: API สำหรับระบบ E-commerce ที่พัฒนาด้วย Go Fiber
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Go E-commerce API
  version: 1.0.0
paths:
...
  /orders:
    get:
      consumes:
      - application/json
      description: ดูคำสั่งซื้อของผู้ใช้ปัจจุบัน
      parameters:
      - default: 1
        description: หน้าที่ต้องการ
        in: query
        name: page
        type: integer
      - default: 10
        description: จำนวนรายการต่อหน้า
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/entities.ApiResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/entities.Order'
                  type: array
                pagination:
                  $ref: '#/definitions/entities.PaginationResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/entities.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ApiResponse'
      security:
      - BearerAuth: []
      summary: ดูคำสั่งซื้อของผู้ใช้
      tags:
      - Orders
    post:
      consumes:
      - application/json
      description: สร้างคำสั่งซื้อใหม่จากตะกร้าสินค้า
      parameters:
      - description: ข้อมูลการสร้างคำสั่งซื้อ
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/entities.CreateOrderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/entities.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/entities.Order'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entities.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/entities.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ApiResponse'
      security:
      - BearerAuth: []
      summary: สร้างคำสั่งซื้อ
      tags:
      - Orders
  /orders/{id}:
    get:
      consumes:
      - application/json
      description: ดูรายละเอียดคำสั่งซื้อตาม ID
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/entities.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/entities.Order'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entities.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/entities.ApiResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/entities.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ApiResponse'
      security:
      - BearerAuth: []
      summary: ดูคำสั่งซื้อตาม ID
      tags:
      - Orders
  /orders/{id}/cancel:
    put:
      consumes:
      - application/json
      description: ยกเลิกคำสั่งซื้อ
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entities.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/entities.ApiResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/entities.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ApiResponse'
      security:
      - BearerAuth: []
      summary: ยกเลิกคำสั่งซื้อ
      tags:
      - Orders
  /orders/admin:
    get:
      consumes:
      - application/json
      description: ดูคำสั่งซื้อทั้งหมดพร้อม pagination (เฉพาะ Admin)
      parameters:
      - default: 1
        description: หน้าที่ต้องการ
        in: query
        name: page
        type: integer
      - default: 10
        description: จำนวนรายการต่อหน้า
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/entities.ApiResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/entities.Order'
                  type: array
                pagination:
                  $ref: '#/definitions/entities.PaginationResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/entities.ApiResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/entities.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ApiResponse'
      security:
      - BearerAuth: []
      summary: ดูคำสั่งซื้อทั้งหมด (Admin)
      tags:
      - Orders
  /orders/admin/{id}/status:
    put:
      consumes:
      - application/json
      description: อัพเดทสถานะคำสั่งซื้อ (เฉพาะ Admin)
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      - description: ข้อมูลการอัพเดทสถานะคำสั่งซื้อ
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/entities.UpdateOrderStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entities.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/entities.ApiResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/entities.ApiResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/entities.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ApiResponse'
      security:
      - BearerAuth: []
      summary: อัพเดทสถานะคำสั่งซื้อ (Admin)
      tags:
      - Orders
  /payments:
    post:
      consumes:
      - application/json
      description: สร้างการชำระเงินสำหรับคำสั่งซื้อ
      parameters:
      - description: ข้อมูลการสร้างการชำระเงิน
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/entities.CreatePaymentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/entities.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/entities.Transaction'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entities.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/entities.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ApiResponse'
      security:
      - BearerAuth: []
      summary: สร้างการชำระเงิน
      tags:
      - Payments
  /payments/{id}/cancel:
    put:
      consumes:
      - application/json
      description: ยกเลิกการชำระเงิน
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entities.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/entities.ApiResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/entities.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ApiResponse'
      security:
      - BearerAuth: []
      summary: ยกเลิกการชำระเงิน
      tags:
      - Payments
  /payments/{id}/verify:
    post:
      consumes:
      - application/json
      description: ยืนยันการชำระเงิน
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: string
      - description: ข้อมูลการยืนยันการชำระเงิน
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/entities.VerifyPaymentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entities.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/entities.ApiResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/entities.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ApiResponse'
      security:
      - BearerAuth: []
      summary: ยืนยันการชำระเงิน
      tags:
      - Payments
...
securityDefinitions:
  BearerAuth:
    description: 'JWT token สำหรับการยืนยันตัวตน ให้ใส่ token ในรูปแบบ: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"